kivyBack/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.1
mysqlclient==2.2.1
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.2
